# Speed up audb.Dependencies.__str__()