--- a/tests/test_load.py
+++ b/tests/test_load.py
@@ -350,7 +350,6 @@ def test_load(dbs, format, version, only_metadata):
 
     # Assert all files are listed in dependency table
     deps = audb.dependencies(DB_NAME, version=version)
-    assert str(deps().to_string()) == str(deps)
     assert len(deps) == (
         len(db.files) + len(db.tables) + len(db.misc_tables) + len(db.attachments)
     )